+++ geotext/geotext.py
@@ -5,7 +5,7 @@ from collections import namedtuple, Counter, OrderedDict
 
 from resources import (
     get_nationalities_data, get_countries_data, get_cities_data,
-    get_words_counts,
+    get_words_counts, get_cities_abbreviations_data, replace_non_ascii,
 )
 
 
@@ -63,6 +63,7 @@ class GeoText(object):
             get_cities_data(min_population=min_population),
             get_countries_data()
         )
+        self._abbreviations = get_cities_abbreviations_data()
         self._update_code_to_country()
         self._location_length = self._get_locations_length()
 
@@ -72,19 +73,33 @@ class GeoText(object):
             words_counts |= get_words_counts(collection)
         return sorted(words_counts, reverse=True)
 
-    def get_candidates(self, text, fuzzy=False):
+    def get_candidates(self, text, fuzzy):
         """
         :param text
         :param fuzzy  if set to True analyze all words in the text regardless
             of capitalization
         """
-        def capitalize_list(l):
-            return map(lambda w: w.title(), l)
-
+        text = replace_non_ascii(text)
         if not fuzzy:
             text = re.sub(r'[^\x00-\x7F]+', ' ', text).strip()
-            return capitalize_list(re.findall(GeoText.LOCATION_REGEX, text))
+            return map(
+                lambda x: x.lower(), re.findall(GeoText.LOCATION_REGEX, text)
+            )
+        # Remove dots from acronyms:
+        text = re.sub(r'\.(?![a-z]{2})', '', text, flags=re.IGNORECASE)
+        # Replace other symbols with spaces
+        # TODO: improve this, since DB has unicode symbols in cities
         text = re.sub(r'[^\w]+', ' ', text).strip()
+
+        # Add abbreviations values: note that we don't replace, but add data
+        # in case abbrev is also a part of location name.
+        result_text = []
+        for word in text.split():
+            result_text.append(word)
+            if word.lower() in self._abbreviations:
+                result_text.append(self._abbreviations[word.lower()].value)
+        text = ' '.join(result_text)
+
         candidates = set()
         for location_length in self._location_length:
             text_substring = text
@@ -94,18 +109,20 @@ class GeoText(object):
                     text_substring
                 ))
                 text_substring = ' '.join(text_substring.split()[1:])
-        return capitalize_list(candidates)
+        return map(lambda x: x.lower(), candidates)
 
-    def read(self, text, fuzzy=False):
-        candidates = self.get_candidates(text, fuzzy=fuzzy)
+    def read(self, text, skip_nationalities=False, fuzzy=True):
+        candidates = self.get_candidates(text, fuzzy)
+        countries = [
+            each for each in candidates if each in self._index.countries
+        ]
         self.countries = [
-            each for each in candidates
-            if each.lower() in self._index.countries
+            self._index.countries[country].original_name
+            for country in countries
         ]
         cities = [
             each for each in candidates
-            if each.lower() in self._index.cities
-            and each.lower() not in self._index.countries
+            if each in self._index.cities and each not in self._index.countries
         ]
         # Iterate over a copy because we'll modify this list
         for city in cities[:]:
@@ -116,24 +133,33 @@ class GeoText(object):
                 if re.findall(r'\b{}\b'.format(city), other_city):
                     cities.remove(city)
                     break
-        self.cities = cities
-        self.nationalities = [
-            each for each in candidates
-            if each.lower() in self._index.nationalities
+        self.cities = [
+            self._index.cities[city].original_name for city in cities
         ]
+        if not skip_nationalities:
+            nationalities = [
+                each for each in candidates
+                if each in self._index.nationalities
+            ]
+            self.nationalities = [
+                self._index.nationalities[nationality].original_name
+                for nationality in nationalities
+            ]
+        else:
+            self.nationalities = nationalities = []
 
         # Calculate number of country mentions
         country_mentions = [
-            self._index.countries[country.lower()]
-            for country in self.countries
+            self._index.countries[country.lower()].value
+            for country in countries
         ]
         country_mentions.extend(
-            [self._index.cities[city.lower()] for city in self.cities]
+            [self._index.cities[city.lower()].value for city in cities]
         )
         country_mentions.extend(
             [
-                self._index.nationalities[nationality.lower()]
-                for nationality in self.nationalities
+                self._index.nationalities[nationality.lower()].value
+                for nationality in nationalities
             ]
         )
         self.country_mentions = OrderedDict(