+++ HISTORY.rst
@@ -3,7 +3,23 @@
 History
 -------
 
-0.1.0 (2016-12-03)
+0.2.0 (2017-02-08)
+------------------
+
+* Handle commas and dots in original location data
+* Add cities abbreviations
+* Add parameter to skip nationalities
+* Handle non-ascii characters in itionaries and input
+* Fuzzy search by default
+
+
+0.1.1 (2016-12-14)
+------------------
+
+* Fix nationalities file
+
+
+0.1.0 (2016-12-06)
 ------------------
 
 * Select cities by minimum population